+++ gemalaya/qml/AppConfigSettings.qml
@@ -2,6 +2,11 @@ import QtQuick 2.14
 import QtQuick.Layouts 1.4
 
 ColumnLayout {
+  Text {
+    text: qsTr('Timers')
+    font.bold: true
+    font.pointSize: 20
+  }
   IntegerCfgSetting {
     dotPath: 'ui.urlCompletionTimeout'
     description: qsTr('URL completion timeout')
@@ -29,7 +34,7 @@ ColumnLayout {
     dotPath: 'ui.page.maximumFlickVelocity'
     description: qsTr('Maximum page flick velocity (in pixels per second)')
     spin.from: 10
-    spin.to: 100000
+    spin.to: 10000
     spin.stepSize: 50
   }
   IntegerCfgSetting {
@@ -43,13 +48,13 @@ ColumnLayout {
     dotPath: 'ui.page.upFlickPPS'
     description: qsTr('How much to flick when "up" is pressed (in pps)')
     spin.from: 10
-    spin.to: 100000
+    spin.to: 10000
     spin.stepSize: 50
   }
   IntegerCfgSetting {
     dotPath: 'ui.page.pageDownFlickPPS'
     description: qsTr('How much to flick when "page down" is pressed (in pps)')
-    spin.from: -1000
+    spin.from: -5000
     spin.to: 0
     spin.stepSize: 50
   }
@@ -57,7 +62,7 @@ ColumnLayout {
     dotPath: 'ui.page.pageUpFlickPPS'
     description: qsTr('How much to flick when "page up" is pressed (in pps)')
     spin.from: 10
-    spin.to: 100000
+    spin.to: 10000
     spin.stepSize: 50
   }
   IntegerCfgSetting {
@@ -79,7 +84,7 @@ ColumnLayout {
     description: qsTr('Font size for links')
     spin.from: 8
     spin.to: 42
-    spin.stepSize: 2
+    spin.stepSize: 1
   }
 
   IntegerCfgSetting {
@@ -87,7 +92,7 @@ ColumnLayout {
     description: qsTr('Font size for normal text')
     spin.from: 8
     spin.to: 42
-    spin.stepSize: 2
+    spin.stepSize: 1
   }
 
   IntegerCfgSetting {
@@ -95,6 +100,6 @@ ColumnLayout {
     description: qsTr('Font size for preformatted text')
     spin.from: 8
     spin.to: 42
-    spin.stepSize: 2
+    spin.stepSize: 1
   }
 }